3/requirements.txt
fastapi==0.110.0
uvicorn==0.15.0
python-jose[cryptography]==3.3.0
bcrypt==3.2.0
pyjwt==2.1.0
python-multipart==0.0.20
httpx==0.28.1
pydantic==2.10.6
psycopg2-binary==2.9.9